Add files via upload
The NN is an example and adventure to express my compression of Neural networks in real life application . Analysing the MNIST plot number analyses. 
Unfortunately the dataset file size was huge so you can better of go to Kaggle and gain access to those files. 
The Kaggle directory will also include test cases.